feat!: move search to home page
What?
- Move search functionality to home page.
- Archive `/search` page.

Why?
- Difficult to implement.
- Difficult to decide which user interaction flow would be more
beneficial for the user -- how user will navigate to search page, etc.
- Current goal is only to migrate to react.
- So, should plan better and work on in it later.
