+++ .archive/src/app/(root)/search/page.tsx
